server: delay transact commit
Changes shouldn't be made to the db unless the we know the whole
transaction will succeed.

This commit changes the Database interface so table indexes can be checked before a transaction is committed.
Once a transaciton is complete, it is commited to the database. This more closely mirrors how the real ovsdb-server works.

In order for this to work, we needed a IndexExists function in the cache.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>